pyproject.toml
[project]
name = "stand-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "codarsssss",email = "i@codarsssss.ru"}
]
#readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
pytest = "^8.3.5"

